refactor: 단순 출력 뷰 패키지 분리
